feat: damage type particles and tooltips

+++ src/main/java/com/aetherteam/aetherii/api/damage/DamageInfliction.java
@@ -5,12 +5,12 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.Item;
 
-public record DamageInfliction(Item item, int slashValue, int impactValue, int pierceValue) {
+public record DamageInfliction(Item item, float slashValue, float impactValue, float pierceValue) {
     public static final Codec<DamageInfliction> CODEC =
             RecordCodecBuilder.create(in -> in.group(
                     BuiltInRegistries.ITEM.byNameCodec().fieldOf("item").forGetter(DamageInfliction::item),
-                    Codec.INT.fieldOf("slash_value").forGetter(DamageInfliction::slashValue),
-                    Codec.INT.fieldOf("impact_value").forGetter(DamageInfliction::impactValue),
-                    Codec.INT.fieldOf("pierce_value").forGetter(DamageInfliction::pierceValue)
+                    Codec.FLOAT.fieldOf("slash_value").forGetter(DamageInfliction::slashValue),
+                    Codec.FLOAT.fieldOf("impact_value").forGetter(DamageInfliction::impactValue),
+                    Codec.FLOAT.fieldOf("pierce_value").forGetter(DamageInfliction::pierceValue)
             ).apply(in, DamageInfliction::new));
 }
